+++ voxcell/math_utils.py
@@ -4,6 +4,7 @@ import functools
 import math
 
 import numpy as np
+from scipy.spatial.transform import Rotation
 
 
 def gcd(a, b):
@@ -72,22 +73,7 @@ def angles_to_matrices(angles, axis):
     Returns:
         (N, 3, 3) array of rotation matrices.
     """
-    i, j, k = {
-        'x': (1, 2, 0),
-        'y': (0, 2, 1),
-        'z': (0, 1, 2),
-    }[axis]
-
-    cos, sin = np.cos(angles), np.sin(angles)
-
-    result = np.zeros((len(angles), 3, 3))
-    result[:, i, i] = cos
-    result[:, i, j] = sin
-    result[:, j, i] = -sin
-    result[:, j, j] = cos
-    result[:, k, k] = 1.0
-
-    return result
+    return Rotation.from_euler(axis, angles).as_matrix()
 
 
 def normalize(vs):